updated the flowchart
separated the file Scripts.drawio into two files. The first one with the same name now shows the scripts as they appear in the original prioject. The other one, `Intended_Script_Flowchart.drawio`, shows the scripts as I organised them in the translation for Hubei. I propose this structure for our project, and as far as I could test it, it already works.