(hopefully) simplified RunModel
The script now has the control variables (those that control what code is run) at the very beginning, so they can not be missed easily. These controls are checked later an and will throw errors when wrong; I have added some tests, but more may be added. I have also changed the generation of the filenames in which the samples are saved.