added a control variable for the cores to RunModel
the script now contains a variable called use_cores, which can be set to an integer or "all". If it is an integer, that is the number of cores used in sampling (checking is used to assure that not too many cores are specified). If it is set to "all" (default), it will use all cores as specified by the call of getOption("mc.cores", 1L).